backend/requirements.txt
Flask==2.3.4
pymongo==4.4.0
python-dotenv==1.0.0
flask-cors==3.0.10
